move UI integration to separate file
