+++ header/lib/terminal.h
@@ -1,13 +1,6 @@
 #pragma once
 #include "expander.h"
 
-void           WINAPI   CustomizeTerminal();
-static BOOL    WINAPI   SubclassMainWindow(HWND hWndMain, BOOL onUiThread);
-static BOOL    WINAPI   SubclassChartWindows(HWND hWndMain, BOOL onUiThread);
-static LRESULT CALLBACK MainWindowSubclassProc(HWND hWnd, uint msg, WPARAM wParam,LPARAM lParam, UINT_PTR subclassId, DWORD_PTR data);
-static LRESULT CALLBACK ChartWindowSubclassProc(HWND hWnd, uint msg, WPARAM wParam,LPARAM lParam, UINT_PTR subclassId, DWORD_PTR data);
-static LRESULT CALLBACK UiHookProc(int code, WPARAM wParam, LPARAM lParam);
-
 char*        WINAPI FindHistoryDirectoryA(const char* filename, BOOL removeFile);
 HWND         WINAPI FindInputDialogA(ProgramType programType, const char* programName);
 DWORD        WINAPI GetCliOptions();
